Remove incorrect assignment to rval during constructor... newer version of SpiderMonkey actually use rval.

+++ sbbs3/bulkmail.cpp
@@ -2,13 +2,13 @@
 
 /* Synchronet bulk e-mail functions */
 
-/* $Id: bulkmail.cpp,v 1.29 2009/10/25 03:12:13 rswindell Exp $ */
+/* $Id: bulkmail.cpp,v 1.30 2010/03/06 00:13:04 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2009 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2010 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -74,7 +74,7 @@ bool sbbs_t::bulkmail(uchar *ar)
 		return(false); 
 	}
 
-	if((length=filelength(fileno(fp)))<=0) {
+	if((length=(long)filelength(fileno(fp)))<=0) {
 		fclose(fp);
 		return(false);
 	}
